refactor(field): move type def to types.field
